requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
google-genai>=0.8.0
streamlit>=1.33.0
